// packages/src/api/generated/data-security/schemas/securityGradeVO.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Data Security Center API (数据安全中心 API 契约)
 * 企业级数据安全与合规治理接口规范（涵盖数据分级、10级分类目录、敏感特征识别、规则调度、密钥管理、动态/静态脱敏与时效白名单）
 * OpenAPI spec version: 1.0.0
 */

export interface SecurityGradeVO {
  activeFieldsCount?: number;
  boundCategoriesCount?: number;
  colorTag: string;
  description?: string;
  gradeCode: string;
  gradeName: string;
  id: number;
  referencedRulesCount?: number;
  /** 1~100 整数 */
  sensitivityScore: number;
  updatedAt?: string;
  updatedBy?: string;
}